Add image diff model
This is where we store the comparison before two images. Although
image order isn't strictly important for a straight comparison
since the alternative is a collection limited to 2 I thought
it would be less confusing to just arbitrarily call one "before"
and the other "after". This also makes it possible to make such
qualitative statements as "thing moved UP"