Add option for taking full height screenshots
You know what's better than a screen diff? A page diff!

This script option lets each image be taken at "full height". We
temporarily resize the browser to the scroll height of the page,
snap a picture, then reset to the original dimensions.

We'll have to see how this plays with JS heavy sites to know if
wait times (for the page to re-draw) are called for.