Use Api key to load correct project
It's no longer possible to generate free-floating revisions.
Currently they must all be associated to a specific scenario.
For now we're just going to grab the first scenario of project
that the api key belongs to.

This will allow us to get our feature specs working again and
make the API compatible with all of the changes that were made to
support scripted scenarios.