Add Scenario policy
For simplicity we'll just say that only project contributors can
have anything to do with scenarios. The pundit controller hooks
got mad if I tried to authorize scnario actions with the
ProjectPolicy directly. This kind of duplication may or may
not be handy. We'll see in the future whether I do end up with
more fine grained scenario specific rules, or whether I just
end up with duplicates of this policy for many child models.