Restrict project tokens view to contributors
Here I've added a method to the policy which does not have an
associated controller action. Instead it's the name of a partial
for the project controller. I'm experimenting with Pundit and
model relationships.

If these tokens were displayed on a separate page then we would
have clear guidance to use the ProjectTokenPolicy but I don't
want that kind of choice to dictate the frontend.