Authorize creating new Comparisons
There's one gotcha here. Since Pundit encourages us to use the
record in question for autorizing we have to be careful we
don't accidentally create that record, return an error message,
but forget to cleanup the illegal record. An alternative would
be to refactor the RevisionCompare to output unsaved records,
authorize then save.